% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_container.R
\name{create_container}
\alias{create_container}
\title{Realiza a criação de um container no Azure Blob Storage}
\usage{
create_container()
}
\description{
Realiza a criação de um container no Azure Blob Storage
}
\details{
Utilize essa função para a criação de um novo container no Azure Blob Storage.
O nome do container será o nome do projeto, utilizando a variavel de ambiente PROJECT_NAME,
definida durante a execução da função generate_r_envirion()

É recomendável a criação de dois containers, um para DEV e outro
para PROD Porém é opcional.


Package necessário: AzureStor
}
\author{
4intelligence
}
