% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_file_on_container.R
\name{delete_file_on_container}
\alias{delete_file_on_container}
\title{Realiza a exclus찾o de um forecast pack Azure Blob Storage.}
\usage{
delete_file_on_container(name)
}
\arguments{
\item{name}{nome do arquivo que deve ser excluido}
}
\description{
Realiza a exclus찾o de um forecast pack Azure Blob Storage.
}
\details{
Realiza a exclus찾o de um forecast pack indicado pelo par창metro name no banco AzureStor.
}
\examples{
\dontrun {
# Excluindo um arquivo do blob
delete_file_on_container('com_co')

}

}
\author{
4intelligence
}
