% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_on_container.R
\name{files_on_container}
\alias{files_on_container}
\title{Realiza a listagem de todos forecast pack contidos no blob Storage.}
\usage{
files_on_container()
}
\value{
dataframe com o nome, tamanho e variável (y) dos arquivos
}
\description{
Realiza a listagem de todos forecast pack contidos no blob Storage.
}
\details{
Realiza a listagem de todos forecast pack contidos no blob Storage do projeto.
}
\examples{
\dontrun{
# Listando todos os arquivos do projeto contidos no blob
files_on_container()
}

}
\author{
4intelligence
}
