% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_r_environ.R
\name{generate_r_environ}
\alias{generate_r_environ}
\title{Gera um arquivo para armazenar as variáveis de autenticação}
\usage{
generate_r_environ(endpoint, access_key, project_name)
}
\arguments{
\item{endpoint}{String com a url base de acesso ao servidor}

\item{access_key}{String com a senha de acesso do usuário}

\item{project_name}{String com o nome do container}
}
\value{
Não existe nenhum retorno para a função.
}
\description{
Função para criar ou adicionar linhas (caso já exista) no arquivo **.Renviron** com os dados fornecidos
}
\details{
O arquivo será gerado com uma única seção, login, que é por padrão a que todas funções neste pacote utilizam.
O local de criação será o mesmo de sua R Session, e irá conter o nome **.Renviron** .

Para acessar o valor das credenciais, utilize a função `Sys.getenv()` e passe como parâmetro a variável que deseja.

Os valores disponíveis são:
`Sys.getenv(AZURE_URL)`
`Sys.getenv(ACCESS_KEY)`
`Sys.getenv(PROJECT_NAME)`
}
\examples{
generate_r_environ("https://4intelligence.com.br/", "example@4i.com.br", "dash-de")

}
\author{
Samuel Souza
}
