% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blob.R
\name{get_blob}
\alias{get_blob}
\title{Realiza o download de algum arquivo armazenado no blob storage}
\usage{
get_blob(blob_name, dest = paste0(getwd(), "/download_data/"))
}
\arguments{
\item{blob_name}{nome do arquivo ou regex para buscar vários arquivos.}

\item{dest}{pasta a ser armazenado os arquivos baixados. Deve ser um caminho para uma pasta.}
}
\value{
Não existe retorno para esta função
}
\description{
Realiza o download de algum arquivo armazenado no blob storage
}
\details{
Realiza o download de um(ns) arquivo(s) em específico, dado um nome ou uma regex.
}
\examples{
\dontrun {
# Baixando um arquivo em específico
get_blob('com_co')

# Baixando vários arquivos, utilizando uma expressão regular
get_blob('^com_')    # Arquivos que começam com o prefixo `com_`.
get_blob('.rds$')    # Arquivos em formato `.rds`.
get_blob('forecast') # Arquivos que contém `forecast` em seu nome.
}

}
\author{
4intelligence
}
