% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_y.R
\name{get_y}
\alias{get_y}
\title{Realiza o download de um forecast pack ou todos os forecast packs}
\usage{
get_y(y = "all", dest = "forecast_packs/")
}
\arguments{
\item{y}{nome da variável dependente. Utilize 'all' para baixar todos os forecast packs de uma vez.}

\item{dest}{nome da pasta de destino do download.}
}
\value{
forecast pack da variável dependente y ou NULL caso todos os forecast packs forem baixados.
}
\description{
Realiza o download de um forecast pack ou todos os forecast packs
}
\details{
Realiza o download de um forecast pack, dado o nome de sua variável dependente y.
Baixa apenas os arquivos que começem com o prefixo 'forecast_' na
pasta \code{dest}.
}
\examples{
\dontrun{
# Baixando o forecast pack de uma variável dependente
forecast_com_co <- get_y(y = 'com_co')
# Visualização dos dados (Opcional)
dplyr::glimpse(forecast_com_co)

# Baixando todos os forecast packs
get_y()

}

}
\author{
4intelligence
}
