% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packs.R
\name{update_packs}
\alias{update_packs}
\title{Realiza o upload de um forecast pack ou todos os forecast packs no AzureStor}
\usage{
update_packs(pack = NULL, path = "packs/")
}
\arguments{
\item{pack}{nome do arquivo a se fazer o upload.

 `NULL` faz o upload de todos os arquivos do caminho passado em `path`.}

\item{path}{pasta ou caminho do arquivo utilizado no parâmetro `pack`.}
}
\description{
Realiza o upload de um forecast pack ou todos os forecast packs no AzureStor
}
\details{
Realiza o upload do forecast pack de cada variavel y no banco AzureStor.
}
\examples{
\dontrun{
# Subindo todos os arquivos da pasta pack para o azuere
 update_packs()

 #Subindo um arquivo da pasta pack para o azuere
 update_packs('com_co')
}
}
\author{
4intelligence
}
